fix: address second-round code review findings (#2347)
Adds missing Gemini finish reasons (IMAGE_PROHIBITED_CONTENT, NO_IMAGE),
fixes silently-passing property-test guards, backtick import matching and
hard-error on unreadable files in the inventory ratchet, real cycle path in
the schema cycle test, non-optional tool_response result default, defensive
usage copies, a barrel-collision guard test, and assorted test strengthening.

+++ packages/core/src/llm-types/tokensAndEmbeddings.test.ts
@@ -107,7 +107,7 @@ describe('tokensAndEmbeddings property-based', () => {
 
   it.prop([
     fc.array(
-      fc.array(fc.float({ min: -1, max: 1, noNaN: true }), {
+      fc.array(fc.double({ min: -1, max: 1, noNaN: true }), {
         minLength: 1,
         maxLength: 10,
       }),
